fix: encode timestamps as CBOR integers to fix Java SDK v2 parsing (#141)
The AWS Java SDK v2 reads CBOR numbers via Double.toString() which
produces scientific notation (e.g. "1.773966938E9") for large floats.
StringToInstant then calls Long.parseLong() which rejects scientific
notation, causing SdkClientException: "Unable to parse date".

Two changes fix this:
- Truncate timestamps to whole epoch seconds (matching real AWS behavior)
  instead of preserving millisecond fractional parts
- In json_to_cbor_impl, convert integer-valued f64 to CBOR Integer
  so the SDK reads them as longs (no scientific notation)
- Add serialize_epoch_seconds custom serializer so JSON responses also
  emit integer timestamps, keeping JSON/CBOR equivalence

+++ src/actions/describe_stream_summary.rs
@@ -33,7 +33,7 @@ pub async fn execute(store: &Store, data: Value) -> Result<Option<Value>, Kinesi
             "StreamARN": stream.stream_arn,
             "StreamName": stream.stream_name,
             "StreamStatus": stream.stream_status,
-            "StreamCreationTimestamp": stream.stream_creation_timestamp,
+            "StreamCreationTimestamp": stream.stream_creation_timestamp as i64,
             "StreamModeDetails": stream.stream_mode_details,
             "OpenShardCount": open_shard_count,
             "ConsumerCount": consumer_count,
